Update furniture and terrain, plastic windows.
Mainly bashing sounds to the new "furniture" and "terrain" from recent PRs.

I saw that pane glass didnt have anything for it but there's so much shit to it that I'll handle it in another pr or commit.

I think i did the open and closing stuff right for the plastic windows.